+++ human_aware_rl/imitation/behavior_cloning_tf2.py
@@ -2,19 +2,18 @@ import os, pickle, copy, shutil, itertools
 from tensorflow import keras
 import tensorflow as tf
 import numpy as np
+from human_aware_rl.rllib.meta_policies import AbstractOffDistrubutionPolicy
 from tensorflow.compat.v1.keras.backend import set_session, get_session
-from human_aware_rl.human.process_dataframes import get_trajs_from_data, get_human_human_trajectories
-from human_aware_rl.static import *
+from human_aware_rl.human.process_dataframes import get_human_human_trajectories
+from human_aware_rl.static import CLEAN_2019_HUMAN_DATA_TRAIN, HUMAN_DATA_DIR
 from human_aware_rl.rllib.rllib import RlLibAgent, softmax, evaluate, get_base_ae, load_trainer
 from human_aware_rl.rllib.policies import StaticPolicy
 from human_aware_rl.rllib.policies import UniformPolicy as DummyOptPolicy
 from human_aware_rl.rllib.utils import get_base_env
 from human_aware_rl.data_dir import DATA_DIR
-from human_aware_rl.utils import recursive_dict_update, get_flattened_keys, create_dir_if_not_exists
+from human_aware_rl.utils import recursive_dict_update, get_flattened_keys
 from overcooked_ai_py.mdp.actions import Action
 from overcooked_ai_py.mdp.overcooked_env import DEFAULT_ENV_PARAMS
-from ray.rllib.policy import Policy as RllibPolicy
-from ray.rllib.models.modelv2 import restore_original_dimensions
 
 #################
 # Configuration #
@@ -545,116 +544,7 @@ class BehaviorCloningPolicy(StaticPolicy):
         if self.eager:
             return NullContextManager()
         return TfContextManager(self._sess)
-
-class AbstractOffDistrubutionPolicy(StaticPolicy):
-
-    """
-    Abstract class for a OOD policy. At a high level, this can be viewed as a meta agent that
-    has both an "on distribution" policy, and an "off distriubtion" policy that 
-    is invoked whenever `self._of_distribution` is true
-
-    Methods:
-        compute_actions:            overrides base RllibPolicy's compute_actions
-        
-    Abstract Methods
-        _on_distribution_init:      Used to initialize on distribution policy. Must return a RllibPolicy type
-        _off_distribution_init:     Used to initialize off distribution policy. Must return a RllibPolicy type
-        _off_distribution:           Invoked to determine whether state is on/off distribution. Returns boolean mask
-
-    Instances:
-        observation_space (gym.spaces.Dict):    Information about observation space. Contains keys "off_dist" and "on_dist"
-        action_space (gym.space.Discrete):      Information about action space
-        on_distribution_policy (RllibPolicy):   Queried for on distribution actions
-        off_distribution_policy (RllibPolicy):  Queried for off-distribution actions
-    """
-
-    def __init__(self, observation_space, action_space, config):
-        """
-        RLLib compatible constructor for initializing a OOD model
-
-        observation_space (gym.Space)           Shape of the featurized observations
-        action_space (gym.space)                Shape of the action space (len(Action.All_ACTIONS),)
-        config (dict)                           Dictionary of relavant policy params
-            - on_dist_config (dict):                Parameters passed into the "on distribution" policy
-            - off_dist_config (dict):               Parameters passed into the "off distrubtion" policy
-        """
-        super(AbstractOffDistrubutionPolicy, self).__init__(observation_space, action_space, config)
-        self.on_distrubtion_policy = self._on_dist_init(config['on_dist_config'])
-        self.off_distrubtion_policy = self._off_dist_init(config['off_dist_config'])
-        
-    def _on_dist_init(self, config):
-        """
-        Initializes on-distribution policy. Must return an RllibPolicy instance
-        """
-        raise NotImplementedError("Must subclass and override this method")
-
-    def _off_dist_init(self, config):
-        """
-        Initializes off-distribution policy. Must return an RllibPolicy instance
-        """
-        raise NotImplementedError("Must subclass and override this method")
-
-
-    def _off_distribution(self, obs_batch, *args, **kwargs):
-        """
-        Determine whether given states are on/off distribution
-
-        Arguments:
-            - obs_batch (np.array) (N, *obs_space.shape): Array of observations. 
-                    Must also be able to handle single, non-batched observation
-
-        Returns
-            - mask (np.array) (N,): Array of booleans, True if off distribution, false otherwise
-        """
-        raise NotImplementedError("Must subclass and override this method")
-
-    def parse_observations(self, obs_batch):
-        # Parse out the on/off-distribution actions
-        if type(obs_batch) == dict:
-            # We received non-batched, non-flattened dictionary observation
-            off_dist_obs, on_dist_obs = obs_batch['off_dist'], obs_batch['on_dist']
-        if hasattr(obs_batch, '__iter__') and type(obs_batch[0]) == dict:
-            # We received batched, non-flattened dictionary observation
-            off_dist_obs, on_dist_obs = [obs['off_dist'] for obs in obs_batch], [obs['on_dist'] for obs in obs_batch]
-        else:
-            # We received batched/unbatched flattened dictionary observation (this is done by default on the Rllib backened
-            # with no public API to disable...). We thus "unpack" the observation to restore it to its original dictionary config
-            restored_batch = restore_original_dimensions(obs_batch, self.observation_space, tensorlib=np)
-            off_dist_obs, on_dist_obs = restored_batch['off_dist'], restored_batch['on_dist']
-
-        return off_dist_obs, on_dist_obs
-
-    def compute_actions(self, obs_batch, *args, **kwargs):
-        off_dist_mask = self._off_distribution(obs_batch, *args, **kwargs)
-        return self._compute_actions(obs_batch, off_dist_mask)
-
-    def _compute_actions(self, obs_batch, off_dist_mask):
-        """
-        Note: Both off- and on-distribution policies are queried for every timestep. There is no lazy computation
-        """
-        off_dist_obs, on_dist_obs = self.parse_observations(obs_batch)
-
-        # On distrubion forward
-        on_dist_actions, on_dist_logits = self._on_dist_compute_actions(on_dist_obs)
-
-        # Off distribution forward
-        off_dist_actions, off_dist_logits = self._off_dist_compute_actions(off_dist_obs)
-
-        # Batched ternary switch based on previously computed masks
-        actions, logits = np.where(off_dist_mask, off_dist_actions, on_dist_actions), np.where(off_dist_mask, off_dist_logits.T, on_dist_logits.T).T
-
-        return actions, [], { "action_dist_inputs" : logits, "off_dist_mask" : off_dist_mask }
-
-    def _on_dist_compute_actions(self, on_dist_obs_batch, **kwargs):
-        actions, _, infos = self.on_distrubtion_policy.compute_actions(on_dist_obs_batch)
-        logits = infos['action_dist_inputs']
-        return actions, logits
-
-    def _off_dist_compute_actions(self, off_dist_obs_batch, **kwargs):
-        actions, _, infos = self.off_distrubtion_policy.compute_actions(off_dist_obs_batch)
-        logits = infos['action_dist_inputs']
-        return actions, logits
-    
+  
 class AbstractBCSelfPlayOPTPolicy(AbstractOffDistrubutionPolicy):
 
     """
